Fix profile loading animation

+++ src/components/profile/Recovery/RecoveryContainer.js
@@ -3,13 +3,13 @@ import { useSelector, useDispatch } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import styled from 'styled-components';
 import RecoveryMethod from './RecoveryMethod';
-import { Translate } from 'react-localize-redux';
 import {
     deleteRecoveryMethod,
     loadRecoveryMethods
 } from '../../../actions/account';
 import SkeletonLoading from '../../common/SkeletonLoading';
 import { useRecoveryMethods } from '../../../hooks/recoveryMethods';
+import { actionsPending } from '../../../utils/alerts'
 
 const Container = styled.div`
 
@@ -37,13 +37,12 @@ const RecoveryContainer = ({ type }) => {
     const [deletingMethod, setDeletingMethod] = useState('');
     const dispatch = useDispatch();
     const account = useSelector(({ account }) => account);
-    const status = useSelector(({ status }) => status);
     const allKinds = ['email', 'phone', 'phrase'];
     const activeMethods = useRecoveryMethods(account.accountId).filter(({ kind }) => allKinds.includes(kind));
     const currentActiveKinds = new Set(activeMethods.map(method => method.kind));
     const missingKinds = allKinds.filter(kind => !currentActiveKinds.has(kind))
     const deleteAllowed = [...currentActiveKinds].length > 1 || account.ledgerKey;
-
+    const recoveryLoader = actionsPending('LOAD_RECOVERY_METHODS')
     missingKinds.forEach(kind => activeMethods.push({kind: kind}));
 
     const handleDeleteMethod = async (method) => {
@@ -56,24 +55,29 @@ const RecoveryContainer = ({ type }) => {
         dispatch(loadRecoveryMethods())
     }
 
-    return (
-        <Container className='recovery-option'>
-            {!status.mainLoader && activeMethods.filter(method => method.kind === type).map((method, i) =>
-                <RecoveryMethod
-                    key={i}
-                    method={method}
-                    accountId={account.accountId}
-                    deletingMethod={deletingMethod === method.publicKey}
-                    onDelete={() => handleDeleteMethod(method)}
-                    deleteAllowed={deleteAllowed}
-                />
-            )}
+    if (!recoveryLoader) {
+        return (
+            <Container className='recovery-option'>
+                {activeMethods.filter(method => method.kind === type).map((method, i) =>
+                    <RecoveryMethod
+                        key={i}
+                        method={method}
+                        accountId={account.accountId}
+                        deletingMethod={deletingMethod === method.publicKey}
+                        onDelete={() => handleDeleteMethod(method)}
+                        deleteAllowed={deleteAllowed}
+                    />
+                )}
+            </Container>
+        )
+    } else {
+        return (
             <SkeletonLoading
-                height='50px'
-                show={status.mainLoader}
+                height='80px'
+                show={recoveryLoader}
             />
-        </Container>
-    );
+        )
+    }
 }
 
 export default withRouter(RecoveryContainer);
